+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## version 0.3.0 (2024/11/13)
+
+- `reporting.icaltools`: some tools using a less sensitive package icalendar
+- new command line `icalcheck <file>`
+- new command line `icalclean <file>`
+
 ## version 0.2.0 (2024/08/30)
 
 - `reporting.Hyperplanning` class able to read ical and make worksheets with Hyperplanning calendars